test: consolidate shared fixtures and drop duplicate specs
* stubAsyncIterator (byte-identical in values.spec.js and yield.spec.js),
  fromArray (identical in errors.spec.js and errors-fail-fast.spec.js)
  and the four open-coded AggregateError-unwrap expressions move to
  test/utils.js — the documented home for shared test helpers — as
  stubAsyncIterator, fromArray and unwrapCapturedError.
* Three specs that pinned behavior already pinned elsewhere are dropped:
  throw.spec.js's signal.aborted-after-throw() (owned by
  per-task-signal.spec.js), errors-fail-fast.spec.js's default-mode
  AggregateError check (errors.spec.js pins the same fixture with
  stronger assertions, also on the default mode), and abort.spec.js's
  fresh-next-after-abort (literally the first iteration of the
  exactly-once spec below it).
* The cleanupTimeout timer-leak assertion becomes a before/after
  countTimers delta, pinning 'close leaves no timer behind' instead of
  'nothing else in the process owns a timer right now'.
* test/utils.js joins the type-coverage ignore list (test/**/*.js) —
  same rationale as the spec files; it is still tsc-strict checked.
